cmd/mp3-slice/mp3lib/copy-with-cancel_test.go
package mp3lib

import (
	"bytes"
	"context"
	"testing"
	"time"
)

// Test the basic functionality with regular-sized sources and destinations.
func TestCopyWithCancel_BasicFunctionalityRegularSize(t *testing.T) {
	srcData := []byte("hello world")
	src := bytes.NewReader(srcData)
	dst := &bytes.Buffer{}

	ctx := context.Background()
	_, err := CopyWithCancel(ctx, dst, src)
	if err != nil {
		t.Fatalf("unexpected error: %v", err)
	}

	if !bytes.Equal(dst.Bytes(), srcData) {
		t.Fatalf("expected %q, got %q", srcData, dst.Bytes())
	}
}

// Test the basic functionality with large sources and destinations.
func TestCopyWithCancel_BasicFunctionalityLargeSize(t *testing.T) {
	srcData := make([]byte, 10*1024*1024) // 10 MB
	for i := range srcData {
		srcData[i] = byte(i % 256)
	}
	src := bytes.NewReader(srcData)
	dst := &bytes.Buffer{}

	ctx := context.Background()
	_, err := CopyWithCancel(ctx, dst, src)
	if err != nil {
		t.Fatalf("unexpected error: %v", err)
	}

	if !bytes.Equal(dst.Bytes(), srcData) {
		t.Fatalf("data mismatch")
	}
}

// Test cancellation before the copying starts.
func TestCopyWithCancel_CancelBeforeStart(t *testing.T) {
	srcData := []byte("hello world")
	src := bytes.NewReader(srcData)
	dst := &bytes.Buffer{}

	ctx, cancel := context.WithCancel(context.Background())
	cancel() // Cancel the context immediately.

	_, err := CopyWithCancel(ctx, dst, src)
	if err == nil {
		t.Fatalf("expected an error due to cancellation, got nil")
	}
	if dst.Len() > 0 {
		t.Fatalf("destination should be empty, but got: %q", dst.Bytes())
	}
}

// Test cancellation in the middle of the copying process.
func TestCopyWithCancel_CancelDuringSlowCopy(t *testing.T) {
	srcData := make([]byte, 10*1024*1024) // 10 MB
	src := bytes.NewReader(srcData)
	// Using a SlowWriter to simulate slower write speeds. Assuming 100ms per kilobyte.
	dst := NewSlowWriter(&bytes.Buffer{}, 10*time.Millisecond)

	ctx, cancel := context.WithCancel(context.Background())
	go func() {
		// Cancel after some delay to simulate cancellation during slow copying.
		time.Sleep(50 * time.Millisecond) // increasing the sleep time due to slow write
		cancel()
	}()

	_, err := CopyWithCancel(ctx, dst, src)
	if err == nil {
		t.Fatalf("expected an error due to cancellation, got nil")
	}
	if dst.w.(*bytes.Buffer).Len() == len(srcData) { // accessing the wrapped buffer to get its length
		t.Fatalf("copy should have been cancelled and not all data copied")
	}
}

// Test copying when the source is empty.
func TestCopyWithCancel_EmptySource(t *testing.T) {
	srcData := []byte{} // Empty data.
	src := bytes.NewReader(srcData)
	dst := &bytes.Buffer{}

	ctx := context.Background()
	_, err := CopyWithCancel(ctx, dst, src)
	if err != nil {
		t.Fatalf("unexpected error: %v", err)
	}
	if dst.Len() > 0 {
		t.Fatalf("destination should be empty, but got: %q", dst.Bytes())
	}
}

// Test a source that returns an error in the middle of the read operation.
func TestCopyWithCancel_SourceErrorMidway(t *testing.T) {}

// Test a destination that returns an error in the middle of the write operation.
func TestCopyWithCancel_DestinationErrorMidway(t *testing.T) {}

// Test both source and destination that return errors.
func TestCopyWithCancel_SourceAndDestinationErrors(t *testing.T) {}

// Test that if the source closes before the destination completes, the copy still completes.
func TestCopyWithCancel_SourceClosesFirst(t *testing.T) {}

// Test that if the destination closes before the source completes, an error is returned.
func TestCopyWithCancel_DestinationClosesFirst(t *testing.T) {}

// Test copying from an empty source to ensure the destination is also empty without errors.
func TestCopyWithCancel_EmptySourceToDestination(t *testing.T) {}

// Test the behavior of the pipe writer closing correctly with the source's error.
func TestCopyWithCancel_PipeWriterCloseWithError(t *testing.T) {}

// Test for scenarios where the pipe might block.
func TestCopyWithCancel_PipeBlockingScenarios(t *testing.T) {}

// Test multiple concurrent calls to CopyWithCancel with different sources and destinations.
func TestCopyWithCancel_MultipleConcurrentCalls(t *testing.T) {}

// Test multiple sequential calls to CopyWithCancel ensuring no effects from previous calls.
func TestCopyWithCancel_MultipleSequentialCalls(t *testing.T) {}

// Test with non-standard IO Readers/Writers.
func TestCopyWithCancel_NonStandardIOBehaviors(t *testing.T) {}

// Test all goroutines within the function terminate correctly.
func TestCopyWithCancel_GoroutineTermination(t *testing.T) {}

// Test there are no resource leaks with io.Pipe's reader and writer.
func TestCopyWithCancel_ResourceLeaks(t *testing.T) {}

// Test if the first error from multiple goroutines is the one propagated by g.Wait().
func TestCopyWithCancel_ErrorGroupFirstErrorPropagation(t *testing.T) {}

// Test both goroutines in the error group return errors simultaneously.
func TestCopyWithCancel_ErrorGroupSimultaneousErrors(t *testing.T) {}
